requirements.txt
gspread==5.12.0
oauth2client==4.1.3
pandas==2.1.4
python-dateutil==2.8.2 